Load default session in absence of session name
The `turses.session.Session.populate` method accepts a
`turses.models.TimelineList` and a session name as a parameter and populates
the timeline list with the timelines defined in the corresponding section of
the sessions file. When no `session` parameter is provided, the default
timelines will be populated.

I've defined the `turses.session.TimelineFactory` class that is able to create
timelines when a name is provided. This class encapsulates the parsing of
timeline names and creation of timelines with the corresponding update
function.`

The `turses.session.clean_timeline_list_string` method accepts a string of
comma-separated timeline names and returns a list of *cleaned* names (i.e. with
the whitespace stripped).

Apart from that, I've written a docstring describing how the sessions are
declared and loaded. The implementation is not finished yet though.